🛠️: refactor `FormDirectory` test to typescript
